Create language_verifier.py
Added language verifier which takes a director (-d), walks its contents looking for .srt files. When it finds one, it assumes the directory it is in is a language code (3 char). It then detects the language using lingua-language-detector. When the languages differ, it deletes the file given -D option. Otherwise it just prints to STDERR.